:tada: feat: sucrose cache
