+++ src/utils.ts
@@ -1,21 +1,7 @@
 import type { BunFile } from 'bun'
-import {
-	Kind,
-	TAnySchema,
-	TModule,
-	TransformKind,
-	type TSchema
-} from '@sinclair/typebox'
-import { Value } from '@sinclair/typebox/value'
-import { TypeCheck, TypeCompiler } from '@sinclair/typebox/compiler'
 
-import { t } from './type-system'
 import type { Sucrose } from './sucrose'
-
-import { mapValueError } from './error'
-
 import type { TraceHandler } from './trace'
-import type { CookieOptions } from './cookies'
 
 import type {
 	LifeCycleStore,
@@ -288,594 +274,6 @@ export const mergeHook = (
 	}
 }
 
-interface ReplaceSchemaTypeOptions {
-	from: TSchema
-	to(options: Object): TSchema | null
-	excludeRoot?: boolean
-	rootOnly?: boolean
-	original?: TAnySchema
-	/**
-	 * Traverse until object is found except root object
-	 **/
-	untilObjectFound?: boolean
-}
-
-export const replaceSchemaType = (
-	schema: TSchema,
-	options: MaybeArray<ReplaceSchemaTypeOptions>,
-	root = true
-) => {
-	if (!Array.isArray(options)) {
-		options.original = schema
-
-		return _replaceSchemaType(schema, options, root)
-	}
-
-	for (const option of options) {
-		option.original = schema
-
-		schema = _replaceSchemaType(schema, option, root)
-	}
-
-	return schema
-}
-
-const _replaceSchemaType = (
-	schema: TSchema,
-	options: ReplaceSchemaTypeOptions,
-	root = true
-) => {
-	if (!schema) return schema
-	if (options.untilObjectFound && !root && schema.type === 'object')
-		return schema
-
-	const fromSymbol = options.from[Kind]
-
-	if (schema.oneOf) {
-		for (let i = 0; i < schema.oneOf.length; i++)
-			schema.oneOf[i] = _replaceSchemaType(schema.oneOf[i], options, root)
-
-		return schema
-	}
-
-	if (schema.anyOf) {
-		for (let i = 0; i < schema.anyOf.length; i++)
-			schema.anyOf[i] = _replaceSchemaType(schema.anyOf[i], options, root)
-
-		return schema
-	}
-
-	if (schema.allOf) {
-		for (let i = 0; i < schema.allOf.length; i++)
-			schema.allOf[i] = _replaceSchemaType(schema.allOf[i], options, root)
-
-		return schema
-	}
-
-	if (schema.not) return _replaceSchemaType(schema.not, options, root)
-
-	const isRoot = root && !!options.excludeRoot
-
-	if (schema[Kind] === fromSymbol) {
-		// eslint-disable-next-line @typescript-eslint/no-unused-vars
-		const { anyOf, oneOf, allOf, not, properties, items, ...rest } = schema
-		const to = options.to(rest)
-
-		if (!to) return schema
-
-		// If t.Transform is used, we need to re-calculate Encode, Decode
-		let transform
-
-		const composeProperties = (v: TSchema) => {
-			if (properties && v.type === 'object') {
-				const newProperties = <Record<string, unknown>>{}
-				for (const [key, value] of Object.entries(properties))
-					newProperties[key] = _replaceSchemaType(
-						value as TSchema,
-						options,
-						false
-					)
-
-				return {
-					...rest,
-					...v,
-					properties: newProperties
-				}
-			}
-
-			if (items && v.type === 'array')
-				return {
-					...rest,
-					...v,
-					items: _replaceSchemaType(items, options, false)
-				}
-
-			const value = {
-				...rest,
-				...v
-			}
-
-			// Remove required as it's not object
-			delete value['required']
-
-			// Create default value for ObjectString
-			if (
-				properties &&
-				v.type === 'string' &&
-				v.format === 'ObjectString' &&
-				v.default === '{}'
-			) {
-				transform = t.ObjectString(properties, rest)
-				value.default = JSON.stringify(
-					Value.Create(t.Object(properties))
-				)
-				value.properties = properties
-			}
-
-			// Create default value for ArrayString
-			if (
-				items &&
-				v.type === 'string' &&
-				v.format === 'ArrayString' &&
-				v.default === '[]'
-			) {
-				transform = t.ArrayString(items, rest)
-				value.default = JSON.stringify(Value.Create(t.Array(items)))
-				value.items = items
-			}
-
-			return value
-		}
-
-		if (isRoot) {
-			if (properties) {
-				const newProperties = <Record<string, unknown>>{}
-				for (const [key, value] of Object.entries(properties))
-					newProperties[key] = _replaceSchemaType(
-						value as TSchema,
-						options,
-						false
-					)
-
-				return {
-					...rest,
-					properties: newProperties
-				}
-			} else if (items?.map)
-				return {
-					...rest,
-					items: items.map((v: TSchema) =>
-						_replaceSchemaType(v, options, false)
-					)
-				}
-
-			return rest
-		}
-
-		if (to.anyOf)
-			for (let i = 0; i < to.anyOf.length; i++)
-				to.anyOf[i] = composeProperties(to.anyOf[i])
-		else if (to.oneOf)
-			for (let i = 0; i < to.oneOf.length; i++)
-				to.oneOf[i] = composeProperties(to.oneOf[i])
-		else if (to.allOf)
-			for (let i = 0; i < to.allOf.length; i++)
-				to.allOf[i] = composeProperties(to.allOf[i])
-		else if (to.not) to.not = composeProperties(to.not)
-
-		if (transform) to[TransformKind as any] = transform[TransformKind]
-
-		if (to.anyOf || to.oneOf || to.allOf || to.not) return to
-
-		if (properties) {
-			const newProperties = <Record<string, unknown>>{}
-			for (const [key, value] of Object.entries(properties))
-				newProperties[key] = _replaceSchemaType(
-					value as TSchema,
-					options,
-					false
-				)
-
-			return {
-				...rest,
-				...to,
-				properties: newProperties
-			}
-		} else if (items?.map)
-			return {
-				...rest,
-				...to,
-				items: items.map((v: TSchema) =>
-					_replaceSchemaType(v, options, false)
-				)
-			}
-
-		return {
-			...rest,
-			...to
-		}
-	}
-
-	const properties = schema?.properties as Record<string, TSchema>
-
-	if (properties && root && options.rootOnly !== true)
-		for (const [key, value] of Object.entries(properties)) {
-			switch (value[Kind]) {
-				case fromSymbol:
-					// eslint-disable-next-line @typescript-eslint/no-unused-vars
-					const { anyOf, oneOf, allOf, not, type, ...rest } = value
-					const to = options.to(rest)
-
-					if (!to) return schema
-
-					if (to.anyOf)
-						for (let i = 0; i < to.anyOf.length; i++)
-							to.anyOf[i] = { ...rest, ...to.anyOf[i] }
-					else if (to.oneOf)
-						for (let i = 0; i < to.oneOf.length; i++)
-							to.oneOf[i] = { ...rest, ...to.oneOf[i] }
-					else if (to.allOf)
-						for (let i = 0; i < to.allOf.length; i++)
-							to.allOf[i] = { ...rest, ...to.allOf[i] }
-					else if (to.not) to.not = { ...rest, ...to.not }
-
-					properties[key] = {
-						...rest,
-						..._replaceSchemaType(rest, options, false)
-					}
-					break
-
-				case 'Object':
-				case 'Union':
-					properties[key] = _replaceSchemaType(value, options, false)
-					break
-
-				default:
-					if (Array.isArray(value.items)) {
-						for (let i = 0; i < value.items.length; i++) {
-							value.items[i] = _replaceSchemaType(
-								value.items[i],
-								options,
-								false
-							)
-						}
-					} else if (
-						value.anyOf ||
-						value.oneOf ||
-						value.allOf ||
-						value.not
-					)
-						properties[key] = _replaceSchemaType(
-							value,
-							options,
-							false
-						)
-					else if (value.type === 'array') {
-						value.items = _replaceSchemaType(
-							value.items,
-							options,
-							false
-						)
-					}
-
-					break
-			}
-		}
-
-	return schema
-}
-
-const createCleaner = (schema: TAnySchema) => (value: unknown) => {
-	if (typeof value === 'object')
-		try {
-			return Value.Clean(schema, structuredClone(value))
-		} catch {
-			try {
-				return Value.Clean(schema, value)
-			} catch {
-				return value
-			}
-		}
-
-	return value
-}
-
-// const unwrapImport = (schema: TImport<any, any>) => {
-// 	if (
-// 		!schema ||
-// 		!schema.$defs ||
-// 		!schema.$ref ||
-// 		!(schema.$ref in schema.$defs)
-// 	)
-// 		return schema
-
-// 	return schema.$defs[schema.$ref]
-// }
-
-export const getSchemaValidator = <T extends TSchema | string | undefined>(
-	s: T,
-	{
-		models = {},
-		dynamic = false,
-		modules,
-		normalize = false,
-		additionalProperties = false,
-		coerce = false,
-		additionalCoerce = []
-	}: {
-		models?: Record<string, TSchema>
-		modules: TModule<any, any>
-		additionalProperties?: boolean
-		dynamic?: boolean
-		normalize?: boolean
-		coerce?: boolean
-		additionalCoerce?: MaybeArray<ReplaceSchemaTypeOptions>
-	} = {
-		modules: t.Module({})
-	}
-): T extends TSchema ? TypeCheck<TSchema> : undefined => {
-	if (!s) return undefined as any
-
-	let schema: TSchema
-
-	if (typeof s !== 'string') schema = s
-	else {
-		const isArray = s.endsWith('[]')
-		const key = isArray ? s.substring(0, s.length - 2) : s
-
-		schema =
-			(modules as TModule<{}, {}>).Import(key as never) ?? models[key]
-
-		if (isArray) schema = t.Array(schema)
-	}
-
-	if (!schema) return undefined as any
-
-	if (coerce || additionalCoerce) {
-		if (coerce)
-			schema = replaceSchemaType(schema, [
-				{
-					from: t.Ref(''),
-					// @ts-expect-error
-					to: (options) => modules.Import(options['$ref'])
-				},
-				{
-					from: t.Number(),
-					to: (options) => t.Numeric(options),
-					untilObjectFound: true
-				},
-				{
-					from: t.Boolean(),
-					to: (options) => t.BooleanString(options),
-					untilObjectFound: true
-				},
-				...(Array.isArray(additionalCoerce)
-					? additionalCoerce
-					: [additionalCoerce])
-			])
-		else {
-			schema = replaceSchemaType(schema, [
-				{
-					from: t.Ref(''),
-					// @ts-expect-error
-					to: (options) => modules.Import(options['$ref'])
-				},
-				...(Array.isArray(additionalCoerce)
-					? additionalCoerce
-					: [additionalCoerce])
-			])
-		}
-	}
-
-	// @ts-ignore
-	if (schema.type === 'object' && 'additionalProperties' in schema === false)
-		schema.additionalProperties = additionalProperties
-
-	if (dynamic) {
-		const validator = {
-			schema,
-			references: '',
-			checkFunc: () => {},
-			code: '',
-			Check: (value: unknown) => Value.Check(schema, value),
-			Errors: (value: unknown) => Value.Errors(schema, value),
-			Code: () => '',
-			Clean: createCleaner(schema),
-			Decode: (value: unknown) => Value.Decode(schema, value),
-			Encode: (value: unknown) => Value.Encode(schema, value)
-		} as unknown as TypeCheck<TSchema>
-
-		if (normalize && schema.additionalProperties === false)
-			// @ts-ignore
-			validator.Clean = createCleaner(schema)
-
-		// @ts-ignore
-		if (schema.config) {
-			// @ts-ignore
-			validator.config = schema.config
-
-			// @ts-ignore
-			if (validator?.schema?.config)
-				// @ts-ignore
-				delete validator.schema.config
-		}
-
-		// @ts-ignore
-		validator.parse = (v) => {
-			try {
-				return validator.Decode(v)
-			} catch (error) {
-				throw [...validator.Errors(v)].map(mapValueError)
-			}
-		}
-
-		// @ts-ignore
-		validator.safeParse = (v) => {
-			try {
-				return { success: true, data: validator.Decode(v), error: null }
-			} catch (error) {
-				const errors = [...compiled.Errors(v)].map(mapValueError)
-
-				return {
-					success: false,
-					data: null,
-					error: errors[0]?.summary,
-					errors
-				}
-			}
-		}
-
-		return validator as any
-	}
-
-	const compiled = TypeCompiler.Compile(schema, Object.values(models))
-
-	// @ts-expect-error
-	compiled.Clean = createCleaner(schema)
-
-	// @ts-ignore
-	if (schema.config) {
-		// @ts-ignore
-		compiled.config = schema.config
-
-		// @ts-ignore
-		if (compiled?.schema?.config)
-			// @ts-ignore
-			delete compiled.schema.config
-	}
-
-	// @ts-ignore
-	compiled.parse = (v) => {
-		try {
-			return compiled.Decode(v)
-		} catch (error) {
-			throw [...compiled.Errors(v)].map(mapValueError)
-		}
-	}
-
-	// @ts-ignore
-	compiled.safeParse = (v) => {
-		try {
-			return { success: true, data: compiled.Decode(v), error: null }
-		} catch (error) {
-			const errors = [...compiled.Errors(v)].map(mapValueError)
-
-			return {
-				success: false,
-				data: null,
-				error: errors[0]?.summary,
-				errors
-			}
-		}
-	}
-
-	return compiled as any
-}
-
-export const getResponseSchemaValidator = (
-	s: InputSchema['response'] | undefined,
-	{
-		models = {},
-		modules,
-		dynamic = false,
-		normalize = false,
-		additionalProperties = false
-	}: {
-		modules: TModule<any, any>
-		models?: Record<string, TSchema>
-		additionalProperties?: boolean
-		dynamic?: boolean
-		normalize?: boolean
-	}
-): Record<number, TypeCheck<any>> | undefined => {
-	if (!s) return
-
-	let maybeSchemaOrRecord: TSchema | Record<number, string | TSchema>
-
-	if (typeof s !== 'string') maybeSchemaOrRecord = s
-	else {
-		const isArray = s.endsWith('[]')
-		const key = isArray ? s.substring(0, s.length - 2) : s
-
-		maybeSchemaOrRecord =
-			(modules as TModule<{}, {}>).Import(key as never) ?? models[key]
-
-		if (isArray)
-			maybeSchemaOrRecord = t.Array(maybeSchemaOrRecord as TSchema)
-	}
-
-	if (!maybeSchemaOrRecord) return
-
-	const compile = (schema: TSchema, references?: TSchema[]) => {
-		if (dynamic)
-			return {
-				schema,
-				references: '',
-				checkFunc: () => {},
-				code: '',
-				Check: (value: unknown) => Value.Check(schema, value),
-				Errors: (value: unknown) => Value.Errors(schema, value),
-				Code: () => '',
-				Clean: createCleaner(schema),
-				Decode: (value: unknown) => Value.Decode(schema, value),
-				Encode: (value: unknown) => Value.Encode(schema, value)
-			} as unknown as TypeCheck<TSchema>
-
-		const compiledValidator = TypeCompiler.Compile(schema, references)
-
-		if (normalize && schema.additionalProperties === false)
-			// @ts-ignore
-			compiledValidator.Clean = createCleaner(schema)
-
-		return compiledValidator
-	}
-
-	const modelValues = Object.values(models)
-
-	if (Kind in maybeSchemaOrRecord) {
-		if ('additionalProperties' in maybeSchemaOrRecord === false)
-			maybeSchemaOrRecord.additionalProperties = additionalProperties
-
-		return {
-			200: compile(maybeSchemaOrRecord, modelValues)
-		}
-	}
-
-	const record: Record<number, TypeCheck<any>> = {}
-
-	Object.keys(maybeSchemaOrRecord).forEach((status): TSchema | undefined => {
-		const maybeNameOrSchema = maybeSchemaOrRecord[+status]
-
-		if (typeof maybeNameOrSchema === 'string') {
-			if (maybeNameOrSchema in models) {
-				const schema = models[maybeNameOrSchema]
-				schema.type === 'object' &&
-					'additionalProperties' in schema === false
-
-				// Inherits model maybe already compiled
-				record[+status] =
-					Kind in schema ? compile(schema, modelValues) : schema
-			}
-
-			return undefined
-		}
-
-		if (
-			maybeNameOrSchema.type === 'object' &&
-			'additionalProperties' in maybeNameOrSchema === false
-		)
-			maybeNameOrSchema.additionalProperties = additionalProperties
-
-		// Inherits model maybe already compiled
-		record[+status] =
-			Kind in maybeNameOrSchema
-				? compile(maybeNameOrSchema, modelValues)
-				: maybeNameOrSchema
-	})
-
-	return record
-}
-
 const isBun = typeof Bun !== 'undefined'
 const hasHash = isBun && typeof Bun.hash === 'function'
 
@@ -888,94 +286,6 @@ export const checksum = (s: string) => {
 	for (let i = 0; i < s.length; ) h = Math.imul(h ^ s.charCodeAt(i++), 9 ** 9)
 
 	return (h = h ^ (h >>> 9))
-}
-
-let _stringToStructureCoercions: ReplaceSchemaTypeOptions[]
-
-export const stringToStructureCoercions = () => {
-	if (!_stringToStructureCoercions) {
-		_stringToStructureCoercions = [
-			{
-				from: t.Object({}),
-				to: () => t.ObjectString({}),
-				excludeRoot: true
-			},
-			{
-				from: t.Array(t.Any()),
-				to: () => t.ArrayString(t.Any())
-			}
-		] satisfies ReplaceSchemaTypeOptions[]
-	}
-
-	return _stringToStructureCoercions
-}
-
-let _coercePrimitiveRoot: ReplaceSchemaTypeOptions[]
-
-export const coercePrimitiveRoot = () => {
-	if (!_coercePrimitiveRoot)
-		_coercePrimitiveRoot = [
-			{
-				from: t.Number(),
-				to: (options) => t.Numeric(options),
-				rootOnly: true
-			},
-			{
-				from: t.Boolean(),
-				to: (options) => t.BooleanString(options),
-				rootOnly: true
-			}
-		] satisfies ReplaceSchemaTypeOptions[]
-
-	return _coercePrimitiveRoot
-}
-
-export const getCookieValidator = ({
-	validator,
-	modules,
-	defaultConfig = {},
-	config,
-	dynamic,
-	models
-}: {
-	validator: TSchema | string | undefined
-	modules: TModule<any, any>
-	defaultConfig: CookieOptions | undefined
-	config: CookieOptions
-	dynamic: boolean
-	models: Record<string, TSchema> | undefined
-}) => {
-	let cookieValidator = getSchemaValidator(validator, {
-		modules,
-		dynamic,
-		models,
-		additionalProperties: true,
-		coerce: true,
-		additionalCoerce: stringToStructureCoercions()
-	})
-
-	if (isNotEmpty(defaultConfig)) {
-		if (cookieValidator) {
-			// @ts-expect-error private
-			cookieValidator.config = mergeCookie(
-				// @ts-expect-error private
-				cookieValidator.config,
-				config
-			)
-		} else {
-			cookieValidator = getSchemaValidator(t.Cookie({}), {
-				modules,
-				dynamic,
-				models,
-				additionalProperties: true
-			})
-
-			// @ts-expect-error private
-			cookieValidator.config = defaultConfig
-		}
-	}
-
-	return cookieValidator
 }
 
 export const injectChecksum = (
