+++ src/ws/index.ts
@@ -9,6 +9,7 @@ import type {
 
 import type { TSchema } from '@sinclair/typebox'
 import type { TypeCheck } from '../type-system'
+import type { ElysiaTypeCheck } from '../schema'
 
 import type { FlattenResponse, WSParseHandler } from './types'
 import type { MaybeArray, Prettify, RouteSchema } from '../types'
@@ -231,7 +232,7 @@ export const createWSMessageParser = (
 }
 
 export const createHandleWSResponse = (
-	validateResponse: TypeCheck<any> | undefined
+	validateResponse: TypeCheck<any> | ElysiaTypeCheck<any> | undefined
 ) => {
 	const handleWSResponse = (
 		ws: ServerWebSocket<any>,
